Created basic authentication.
Created a database for Slothbucket in MongoDB, as well as a 'users' collection. With this in place (and some other tools), users are able to register, login, view their profile, etc.
To achieve this, I made a lot of changes:
- Added in directories/files for routes, controllers, utils, and models
- Added in mongoose to simplify interactions with MongoDB
- Added in bcrypt for hashing and salting passwords
- Added in express-session and connect-mongo for handling and storing sessions
- Added in helmet for more secure HTTP headers
On top of this, I created the needed pages for basic routes in a user system (login, profile, and register). Also, I attempted to make the logging system more reusable by putting it in its own file and requiring it as needed, rather than having to type out all the winston/logger options every time.