Finished working on request changes.
Created a config.js file to hold all sensitive credentials and added it to be ignored in the .gitignore files. Moved all the DB and authentication logic into the app/ directory and changed some code so it would all work. Continued managing session in the ui/ directory, storing the JWT/user id in the cookie and using MongoDB as the session store. Also, wrote some tests and did some other minor fixes.