Create login system (#13)
* Made a very simple application skeleton to get started.

* Added a .gitignore file.

* Created basic authentication.

Created a database for Slothbucket in MongoDB, as well as a 'users' collection. With this in place (and some other tools), users are able to register, login, view their profile, etc.
To achieve this, I made a lot of changes:
- Added in directories/files for routes, controllers, utils, and models
- Added in mongoose to simplify interactions with MongoDB
- Added in bcrypt for hashing and salting passwords
- Added in express-session and connect-mongo for handling and storing sessions
- Added in helmet for more secure HTTP headers
On top of this, I created the needed pages for basic routes in a user system (login, profile, and register). Also, I attempted to make the logging system more reusable by putting it in its own file and requiring it as needed, rather than having to type out all the winston/logger options every time.

* Finished working on request changes.

Created a config.js file to hold all sensitive credentials and added it to be ignored in the .gitignore files. Moved all the DB and authentication logic into the app/ directory and changed some code so it would all work. Continued managing session in the ui/ directory, storing the JWT/user id in the cookie and using MongoDB as the session store. Also, wrote some tests and did some other minor fixes.

* Added a mock for the config module to the unit test for verify.js

* Switched out config.js for env variables.

* Fixed requested changes, added in tests, and did some minor restructuring and syntax fixes.

* Made a very simple application skeleton to get started. (#12)

* Made a very simple application skeleton to get started.

* Added a .gitignore file.