Created the start of the image uploading system.
Added in a POST route for image uploads, 'upload-image,' and created all the basic functions to get it working. With this route, the user uploads a file, and that file is encoded to base64 and posted to Slothbucket's 'classify-image' route. We take Slothbucket's response to tell the user if the image contains a sloth or not (or handle any errors).