Create image upload (#16)
* Created the start of the image uploading system.

Added in a POST route for image uploads, 'upload-image,' and created all the basic functions to get it working. With this route, the user uploads a file, and that file is encoded to base64 and posted to Slothbucket's 'classify-image' route. We take Slothbucket's response to tell the user if the image contains a sloth or not (or handle any errors).

* Fixed a test in the security controller.

* Rounded out some features for uploading images and made some minor fixes.

Made it so user can view the image right after they upload it (and it's successfully classified as containing a sloth). Also, made some structural changes for better and faster flow. Fixed a small error in the write_file's controller.

* Finished up the basic image uploading system.

Added in functionality for an image library, so that a user can view all of the images they've uploaded. Adding in the image library allowed for redirecting to that page after uploading an image. Also, improved error handling and some other minor fixes, including minor syntax fixes. This required updating some tests, so I went ahead and fixed some old tests, too.

* Removed an unused variable

* Just made a really quick change to the UI so it's a little more structured, and fixed a few useless tags.

* Fixed requested changes and made a few simple UI changes.