Got the route set up for running a demo on images.
Added in a route so the client can send a request with base64 and get back a Tensorflow classification. The route is for demo purposes, so a user does not have to sign up before seeing how Slothbucket works. For this route, I also had to create a module to handle deleting files.

Also, I updated some dependency features that were deprecated.